+++ tests/acceptance/new-home-test.ts
@@ -39,6 +39,20 @@ module('Acceptance | new home page test', hooks => {
         assert.dom('[data-test-arrow]').exists({ count: 3 });
         assert.dom('[data-test-learn-more-button]').exists();
 
+        // Check integrations
+        assert.dom('[data-test-integrations-heading]')
+            .containsText(t('new-home.integrations-section.header').toString());
+        assert.dom('[data-test-authentication-heading]')
+            .containsText(t('new-home.integrations-section.authentication').toString());
+        assert.dom('[data-test-discovery-heading]')
+            .containsText(t('new-home.integrations-section.discovery').toString());
+        assert.dom('[data-test-references-heading]')
+            .containsText(t('new-home.integrations-section.references').toString());
+        assert.dom('[data-test-storage-heading]')
+            .containsText(t('new-home.integrations-section.storage').toString());
+        assert.dom('[data-test-logo]').exists({ count: 8 });
+        assert.dom('[data-test-get-started-button]').exists({ count: 2 });
+
         // Check footer.
         assert.dom('footer').exists();
         await a11yAudit();
@@ -55,6 +69,7 @@ module('Acceptance | new home page test', hooks => {
         assert.dom('[data-test-add-research-heading]').exists();
         assert.dom('[data-test-add-research-subheading]').exists();
         assert.ok(document.querySelector('[data-test-hero-container]')!.className.includes('versionB'));
+        assert.dom('[data-test-get-started-button]').exists({ count: 1 });
 
         await a11yAudit();
         await percySnapshot(assert);